interactive/requirements.txt
bidict==0.22.0
certifi==2022.9.24
charset-normalizer==2.1.1
idna==3.4
printy==2.2.0
python-engineio==4.3.4
python-socketio==5.7.2
requests==2.28.1
urllib3==1.26.12
websocket-client==1.4.2
